feat: 🎸 initialize the "adminFirebaseApp" object
initializes some of the initial SDK files of "firebase-admin"

BREAKING CHANGE: 🧨 the previous configuration has been replaced, all various parts of the
code will break
